requirements.txt
-i https://pypi.org/simple
adal
atomicwrites
attrs
azure-common
azure-core
azure-graphrbac
azure-mgmt-authorization
azure-mgmt-containerregistry
azure-mgmt-core
azure-mgmt-keyvault
azure-mgmt-resource
azure-mgmt-storage
azureml-core
azureml-defaults
backports.tempfile
backports.weakref
certifi
cffi
chardet
click
colorama
configparser
contextlib2
cryptography
docker-pycreds
docker
filelock
flake8
gitdb
gitpython
huggingface-hub
hydra
idna
iniconfig
isodate
isort
jeepney
jmespath
joblib
jsonpickle
matplotlib
mccabe
msrest
msrestazure
ndg-httpsclient
nltk
numpy
oauthlib
packaging
pandas
pathspec
pathtools
pluggy
promise
protobuf
psutil
py
pyasn1
pycodestyle
pycparser
pyflakes
pyjwt
pyopenssl
pyparsing
pytest
python-dateutil
pytz
pywin32; sys_platform == 'win32'
pyyaml
regex
requests-oauthlib
requests
ruamel.yaml.clib; platform_python_implementation == 'CPython' and python_version < '3.10'
ruamel.yaml
sacremoses
scipy
secretstorage
sentry-sdk
shortuuid
six
smmap
subprocess32
tokenizers
toml
torch
tqdm
transformers
typing-extensions
urllib3; sys_platform == 'win32' or sys_platform == 'cygwin'
wandb
websocket-client
wheel
yapf
